server/requirements.txt
anyio==3.6.2
astroid==2.12.13
click==8.1.3
colorama==0.4.6
dill==0.3.6
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
platformdirs==2.5.4
psycopg2-binary==2.9.5
pydantic==1.10.2
pylint==2.15.7
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.22.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
uvicorn==0.20.0
wrapt==1.14.1
git+https://github.com/gim-hangil/korail2.git@master
